Try again to disable core dumps in some tests
133767185047987e1d7f768d09db950dc8e38897 attempted to disable core
dumps in some tests by adding a --disable_core_dumps flag. However,
I had put the handling of this flag in the InitKuduOrDie() function,
which actually is called prior to flags being parsed, so the flag
didn't work as intended.

I noticed that the fix was ineffectual because the tests that were
supposed to be de-flaked by it remained flaky. After setting up
a VM which matched the test environment I noticed that they were
still flaky due to core dumps, and figured out this issue.

The fix is just to move the DisableCoreDumps() call to the spot
where we parse flags.

Change-Id: I51f3d5d8c72622b8a3b8b4a13dd343e864a78db3
Reviewed-on: http://gerrit.cloudera.org:8080/3549
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>