master: handle RPC responses on reactor threads
The introduction of the election rwlock requires that the handling of RPC
responses move off of the singleton thread pool used for invoking the
election callback, otherwise waiting on outstanding RPCs could lead to a
deadlock. Rather than introduce a new thread pool, I observed that none of
the responses take a lock or perform IO, so we can simplify the code a bit
by operating directly on the reactor threads.

There is one exception to the above statement: retried RPCs invoke Run() on
the reactor thread, and Run() can (theoretically) block on DNS resolution.
This has always been the case, though, so maybe we don't care?

Change-Id: I1af5a77ffa66fcc39f71243529a40aae2b542971
Reviewed-on: http://gerrit.cloudera.org:8080/3583
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>