[java-client] Fix a hang in TabletClient
TabletClient#sendRpc has been changing a lot lately, and it's finicky.
In this case we weren't handling RPCs that we needed to retry when the
connection had already been closed before we even tried to encode the
RPC. I've only been able to reproduce this on the cluster with really
slow inserts.

Change-Id: I5f92a2f9464f245182a59178f2eab8a88422864c
Reviewed-on: http://gerrit.cloudera.org:8080/3586
Reviewed-by: Dan Burkert <dan@cloudera.com>
Tested-by: Kudu Jenkins