Add ComputeIfAbsent methods to map-util
This adds two new ComputeIfAbsent methods to map-util, inspired by java's implementation,
which can be found here:

https://docs.oracle.com/javase/8/docs/api/java/util/concurrent/ConcurrentHashMap.html#computeIfAbsent-K-java.util.function.Function-

The first one, ComputeIfAbsent, is a simple c++ implementation of the exact same
concept: Retreive a value from a map; if it doesn't exist compute it using a lambda
function and insert it first.

The second one is a slight variation that returns a pair instead of a pointer to the
value. The pair includes the pointer to the value, but also a bool indicating whether
the value was initially absent.

Change-Id: Iba9fa8038e699e66d34ce541cd02c77f46691315
Reviewed-on: http://gerrit.cloudera.org:8080/3593
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins