[java-client] Re-enable multi-master tests
This patch makes TestMasterFailover useful again. It also adds the killing of masters
to ITClient. Finally, it sets the raft heartbeat lower so that we don't wait 1.5s for
leader elections.

TestGetMasterRegistrationReceived was added since a previous version of this patch
encountered a bug that such a simple unit test can detect.

Change-Id: Ia1051222738c84ef3d3e1a33b4981bc9454b7972
Reviewed-on: http://gerrit.cloudera.org:8080/3654
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>