KUDU-1492: Show column encodings/compression on table page in master
This displays column attributes for the table schema in picture.
If the table schema doesn't specify these attributes,
AUTO_ENCODING or DEFAULT_COMPRESSION are displayed which means
whatever is the current default value attributes in the given
release. E.g., release 0.9.1 has values PLAIN_ENCODING for encoding
and NO_COMPRESSION for compression.
Sample results are posted in JIRA KUDU-1492.

Change-Id: I990a8d790ab71a05be04f0b7468b5da0894478e8
Reviewed-on: http://gerrit.cloudera.org:8080/3667
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>