[java-client] fix soundness hole in flushing async kudu session
This fixes an issue where AsyncKuduSession#apply in AUTO_FLUSH_BACKGROUND mode
could refresh the active buffer and add an initial operation to it, but fail to
set a flush timer on it. Writing a regression test proved to be very difficult
since it relies on a lot of timing variables around how fast batches can flush.

Change-Id: Ic838b65e2ccf5c48f0b51db235936b834aa24cae
Reviewed-on: http://gerrit.cloudera.org:8080/3676
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Jean-Daniel Cryans <jdcryans@apache.org>