C++ client: fix on KuduSession::GetPendingErrors()
If re-using the same std::vector instance as a placeholder
for repeated calls to KuduSession::GetPendingErrors(), irrelevant
error information can be pushed into the session error container.
Clean the placeholder container before swapping it with current
error container of the KuduSession.

Also, make it possible to pass NULL for the 'overflowed' out parameter.

Change-Id: Iafee46a6135ac2096d1f8b690da5b084a15eb73e
Reviewed-on: http://gerrit.cloudera.org:8080/3677
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins