[java client] Fix a race involving the read timeout
The ReadTimeoutHandler isn't something we really want (per-RPC timer
would be better than socket read timeout), but some handling that was
added a year ago (01698c3) made things worse. Being the author of that other
patch, I think my fix was misguided.

The race this patch fixes is that some RPCs were able send their payload
on the wire even after cleanup() had run, but before `chan` was set to
null, when a ReadTimeoutException was fired. The proper way is to close
the channel first, then start cleaning up, so that concurrent threads who
process RPCs won't think that we're all good because the channel isn't null.

I'm easily able to repro this on a cluster with a 1s read timeout along
with ITBLL.

Change-Id: I486fb4c099a9f110f17b4bc72e48d91f258ee044
Reviewed-on: http://gerrit.cloudera.org:8080/3724
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>