[make_site.sh] an option for the script: doxygen
By default, do not build the doxygen docs for the site.
Building auto-generated doxygen docs requires the 'doxygen' tool
to be present at build machines, which is not the case right now.

To build doxygen docs, add 'doxygen' option when running the script.
The latter happens automatically for the 'site' target
when the doxygen package is present at the build machine.

Change-Id: I23394d4339b9afb696335d9795f11d8029999ff9
Reviewed-on: http://gerrit.cloudera.org:8080/3766
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>