external_mini_cluster: adjust single master data directory
If we stored single master data in "master-0" and not "master", it'd be
possible to reuse single master cluster data in a multi master cluster. This
should be a purely cosmetic change, but I've put it in its own patch to draw
more attention to it in case I'm wrong.

I snuck in a change to remove --enable_leader_failure_detection=true as
that's now the default value.

Change-Id: Ia1cfcd529b14cb2472a8673edbc28dbe66175a8a
Reviewed-on: http://gerrit.cloudera.org:8080/3810
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>