Rename Remote Bootstrap to Tablet Copy (part 1)
This does a bunch of automated substitutions of the terminology,
renaming "remote bootstrap" to "tablet copy". This commit does
not rename the files yet, so that it's easier to review without
combining a file name change with the file content change.

This is an incompatible change since some flag and metric names are
changed. Another follow-up will have to add release notes.

Change-Id: Ic017e9121ed365cb3d9c9c81430915bf5b295bf6
Reviewed-on: http://gerrit.cloudera.org:8080/3848
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>