Rename Remote Bootstrap: some manual changes
This patch has a bunch more manual fixes to replace "Remote Bootstrap"
terminology with "Tablet Copy". I looked through the previous (automatic)
diff and tried to clean up comments and variable names to be more clear,
as well as find other abbreviated cases like the term 'bootstrap' or 'rb'
which needed to be updated.

This still doesn't do the file renames. That's yet to come.

Change-Id: I9e2860b64e46c5cc021cb8574e6845d7bcb27cce
Reviewed-on: http://gerrit.cloudera.org:8080/3852
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Kudu Jenkins