Add AvroKuduEventProducer to Kudu-Flume integration
This adds a KuduEventProducer that accepts events with Avro-encoded
bodies and allows the KuduSink to push them to Kudu. It only accepts
schemas of Record type that hold values of types compatible with
Kudu's types. The schema is provided to the EventProducer ahead of
time or in the event header, as a URL or as a literal.

Change-Id: I6715df72e447e72f4801a2e026f6840d09b401e1
Reviewed-on: http://gerrit.cloudera.org:8080/4034
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>