integration_tests: end-to-end master permanent failure test
This commit defines a workflow for handling master permanent failures, and
includes an integration test to execute it. In production, the workflow will
be run manually by an admin, or via a script.

As part of this change, I added support for optional parameters to the new
CLI tool. It was straight-forward to reuse gflags for this.

Still to come: CLI help for positional (i.e. required) parameters.

Change-Id: Ibfab2561e066f00fe56eb4f5d6d6ccbbb2dcbed5
Reviewed-on: http://gerrit.cloudera.org:8080/3969
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>