Create base class for MiniCluster and ExternalMiniCluster
This allows utility classes to interoperate with both of these types of
clusters.

Some minor changes to support this:

* Consistent interface for CreateClient()
* Consistent interface for Shutdown()

Additionally, I removed the 'readability/inheritance' rule from
'make lint' because it prevents using both 'virtual' and 'final' on the
same method, however that syntax is required by Clang in C++11 mode.

Change-Id: I62256168a9245c845e99f7253f07e69a225954bf
Reviewed-on: http://gerrit.cloudera.org:8080/3974
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>