Support downgrade to version that has LocalConsensus
Continue writing the "local=false" field into RaftConfigPB when we have
more than one member of the config.

This partially reverts a few lines of code from revision
35a92b49913ceea80d9c0c61cb7af2ebcf007751, which removed the
LocalConsensus implementation from Kudu. Because the "local" field will
no longer be used, the code writing it was removed. However, since that
field is used by previous versions, and defaults to true if unset, if we
want to allow users to downgrade Kudu to an older version then we need
to keep writing it for a while.

Change-Id: I6d324b4f3ce5906bfc9a4df49dd60ca0cb919820
Reviewed-on: http://gerrit.cloudera.org:8080/3985
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Todd Lipcon <todd@apache.org>