Enable replay cache by default
We've done looping of the stress tests, as well as some significant
cluster testing (YCSB on 72 node cluster for several days). In the
cluster, we didn't see any significant increased memory usage due to
this feature, nor any stability issues.

Additionally, the risk of the feature is relatively low: the worst
it could cause is memory leaks or crashes, but no chance for permanent
data corruption or loss. For the time being, I am leaving the ability
to disable the feature via a gflag, in case we do see any instability.
We can remove the flag entirely in the 1.0 timeframe.

Change-Id: I35b5e74ac30aa3309a0a7e035c8dff7f61c3f275
Reviewed-on: http://gerrit.cloudera.org:8080/4002
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins