Actually support downgrade to version that has LocalConsensus
Commit 74210b2546df9fd5dec7bb926eeb524362d2da90 attempted to support
downgrade from 0.10.x to 0.9.x however there is specific validation in
VerifyRaftConfig() in Kudu 0.9.x that requires the optional field
"local" to be set (checked using has_local()). So we must not rely on
the default, rather, we must always set that field.

This patch was manually tested as follows:

A 3-way replicated table was created in 0.10.0-RC1 plus this patch,
using a single master and 3 tablet servers. The services were killed,
then restarted using 0.9.1 binaries.

Change-Id: Icded76c6400b4802581671e2e41efb74a47c5ec4
Reviewed-on: http://gerrit.cloudera.org:8080/4059
Tested-by: Kudu Jenkins
Reviewed-by: Dinesh Bhat <dinesh@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>